IOS-2943 Revert "TMP Convert to uint16"
This reverts commit 455c0052c141dd18845c96644ab42ade5dcf6432.

+++ pEpForiOS/AppDelegate+PredeployedAccounts.swift
@@ -33,8 +33,8 @@ extension AppDelegate {
                                  imapServerPort: Int,
                                  smtpServerName: String,
                                  smtpServerPort: Int) {
-        let imapServerDict = serverDictionary(name: imapServerName, port: UInt16(imapServerPort))
-        let smtpServerDict = serverDictionary(name: smtpServerName, port: UInt16(smtpServerPort))
+        let imapServerDict = serverDictionary(name: imapServerName, port: imapServerPort)
+        let smtpServerDict = serverDictionary(name: smtpServerName, port: smtpServerPort)
     }
 
     private func serverDictionary(name: String, port: UInt16) -> SettingsDict {
